Refactor to use OpenAPI SDK (#431)
* Refactor to use OpenAPI SDK

* Alerts

* OneOfs renaming

* logs threshold

* Logs Ratio Threshold

* Time relative

* Metric threshold

* Tracing threshold

* Tracing immdeiate

* Logs anomaly

* Metric anomaly

* Logs new value

* Logs Unique Count

* test and prom rule adjustments

* SLO threshold

* [skip ci] webhooks

* dashboards

* dashboard folders

* groups

* use IsNotFound

* archive logs

* e2m

* custom roles

* scopes

* bump sdk

* use openapi sdk for alert internal logic, e.g. listing webhooks

* some fixes

* tco logs and traces

* slo

* update nc

* add some tests

* slo samples

* slo ref in alerts

* update sdk

* update tests

* sync chart

* fix alert slo example

* notification center additions

* use new archive retentions client in tco

* revert dashboard, alertscheduler, e2m

* fix views

* formatting

+++ internal/controller/prometheusrule_controller.go
@@ -452,7 +452,7 @@ func prometheusAlertToMetricThreshold(rule prometheus.Rule, priority coralogixv1
 			},
 		},
 		MissingValues: coralogixv1beta1.MetricMissingValues{
-			MinNonNullValuesPct: ptr.To(uint32(0)),
+			MinNonNullValuesPct: ptr.To(int64(0)),
 		},
 	}
 }
